+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## [Unreleased]
 
+## [1.1.0] - 2018-07-17
+### Added
+- check to see if site is running
+- check to see if app pool is running
+
 ## [1.0.0] - 2017-06-07
 ### Added
 - Support for Ruby 2.3 and 2.4 (@eheydrick)
@@ -25,5 +30,6 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - initial release
 
 [Unreleased]: https://github.com/sensu-plugins/sensu-plugins-iis/compare/1.0.0...HEAD
+[1.1.0]: https://github.com/sensu-plugins/sensu-plugins-iis/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/sensu-plugins/sensu-plugins-iis/compare/0.0.2...1.0.0
 [0.0.2]: https://github.com/sensu-plugins/sensu-plugins-iis/compare/0.0.1...0.0.2